+++ src/api.ts
@@ -1,17 +1,16 @@
-import axios, { AxiosRequestConfig, AxiosResponse, Method } from 'axios';
+import axios, { AxiosRequestConfig, AxiosResponse } from 'axios';
 
 import get from './getter';
 import localStorage from './localStorage';
-import { ApiResponse } from './types/api';
+import { ApiResponse, UrlConfigType } from './types/api';
 
 export function callApi(
-  url: string,
-  method: Method = 'GET',
+  urlConfig: UrlConfigType,
   config: AxiosRequestConfig = {}
 ): Promise<ApiResponse<any>> {
   return axios({
-    method,
-    url,
+    method: urlConfig.method,
+    url: urlConfig.url,
     ...config,
   }).then(
     (result: AxiosResponse<any>) => {
@@ -39,9 +38,8 @@ export function callApi(
 }
 
 /**
- * Метод для отправки запроса к API.
- * @param {string} endpoint URL, на который нужно отправить запрос
- * @param {Method} methodType Метод запроса, по умолчанию 'GET'
+ * Метод для отправки запроса к api.
+ * @param {UrlConfigType} urlConfig URL, на который нужно отправить запрос, вместе с методом запроса
  * @param {any} data Тело запроса либо GET-параметры в виде объекта
  * @param {AxiosRequestConfig} config Конфиг axios
  * @param {boolean} multipartFormData Содержит ли запрос данные формы
@@ -50,8 +48,7 @@ export function callApi(
  * иначе поля error and errorData с информацией об ошибке.
  */
 export default function api(
-  endpoint: string,
-  methodType: Method = 'GET',
+  urlConfig: UrlConfigType,
   data: null | undefined | any = {},
   config: AxiosRequestConfig = {},
   multipartFormData = false,
@@ -61,7 +58,7 @@ export default function api(
 
   if (
     (queryConfig.data === null || queryConfig.data === undefined) &&
-    methodType !== 'GET'
+    urlConfig.method !== 'GET'
   ) {
     queryConfig.data = data;
   }
@@ -95,9 +92,7 @@ export default function api(
     });
   }
 
-  const queryMethodType = methodType.toUpperCase() as Method;
-
-  if (queryMethodType === 'GET') {
+  if (urlConfig.method === 'GET') {
     queryConfig.params = {
       ...data,
       uid: localStorage.getItem('userId'),
@@ -108,7 +103,7 @@ export default function api(
     };
   }
 
-  return callApi(endpoint, queryMethodType, queryConfig)
+  return callApi(urlConfig, queryConfig)
     .then((response) => ({ response }))
     .catch((error) => {
       return {