+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [+] usePolling
 - [+] usePreviousState
 - [+] useValueTransition
+- [+] checkDev
 - [*] в initializeAppParams устанавливается поле is_dev
 - [*] изменена сигнатура метода pluralize, добавлен метод plural
 - [*] в randomNumber добавлена возможность получать дробное число